gatekeeper: rename package
